+++ campagnodon_civicrm.php
@@ -166,6 +166,7 @@ function campagnodon_civicrm_civicrm_alterAPIPermissions($entity, $action, &$par
   $permissions['campagnodon'] = [
     'start' => ['Campagnodon api'],
     'dsp2info' => ['Campagnodon api'],
-    'updatestatus' => ['Campagnodon api']
+    'updatestatus' => ['Campagnodon api'],
+    'convert' => ['Campagnodon api']
   ];
 }